refactor(governance): move member management routes
