+++ docs/context/architecture/multi-tenancy.md
@@ -146,7 +146,7 @@ identity classification. `api.py` re-exports the same objects so existing caller
 retain their import path while route definitions move out incrementally. Session signing and validation
 live beside it in `domain.identity.session`; `treg.session` remains a compatibility alias, while runtime
 owners use the domain path. The former transitional `routers.dependencies` module is deleted.
-`routers.orgs` owns signup, team-entry, and invite-lifecycle HTTP translation. `application.signup` owns the signup sessions,
+`routers.orgs` owns signup, team-entry, invite-lifecycle, and member-management HTTP translation. `application.signup` owns the signup sessions,
 commits, attribution, referral sequencing, and promotional grant; `domain.governance.teams` owns slug,
 Org and Membership creation rules, and the team-switcher read model.
 - **Registration is shared across doors:** `application.signup.find_or_create_user(db, email)` finds a user or creates them
